requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==3.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
motor==3.6.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.40.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.34.2
